requirements.txt
google-genai>=1.0.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=1.12.0

